projects/week-3-doc-search/pyproject.toml
[project]
name = "week-3-doc-search"
version = "0.1.0"
description = "Document search CLI with vector and hybrid search"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "docling>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.0.0",
]
